Partial fix of formatters for new output structure
