+++ packages/checkup-plugin-ember/src/formatters/ember-template-lint-summary-formatter.ts
@@ -1,6 +1,6 @@
-import { FormatterArgs, renderLintingSummaryResult } from '@checkup/core';
+import { BaseOutputWriter, renderLintingSummaryResult } from '@checkup/core';
 import { Result } from 'sarif';
 
-export function format(taskResults: Result[], args: FormatterArgs) {
-  renderLintingSummaryResult(taskResults, args);
+export function format(taskResults: Result[], writer: BaseOutputWriter) {
+  renderLintingSummaryResult(taskResults, writer);
 }
