+++ packages/checkup-plugin-javascript/src/formatters/outdated-dependencies-formatter.ts
@@ -1,9 +1,9 @@
-import { sumOccurrences, renderEmptyResult, FormatterArgs } from '@checkup/core';
+import { sumOccurrences, renderEmptyResult, BaseOutputWriter } from '@checkup/core';
 import { Result } from 'sarif';
 
-export function format(taskResults: Result[], args: FormatterArgs) {
-  args.writer.section(taskResults[0].properties?.taskDisplayName, () => {
-    args.writer.sectionedBar(
+export function format(taskResults: Result[], writer: BaseOutputWriter) {
+  writer.section(taskResults[0].properties?.taskDisplayName, () => {
+    writer.sectionedBar(
       taskResults.map((result: Result) => {
         return result.message.text === 'No results found'
           ? renderEmptyResult(result)
