+++ packages/cli/src/formatters/formatter-utils.ts
@@ -1,63 +1,63 @@
-import { Notification } from 'sarif';
-import { CheckupMetadata, FormatterArgs } from '@checkup/core';
-import { yellow } from 'chalk';
+// import { Notification } from 'sarif';
+// import { CheckupMetadata, FormatterArgs } from '@checkup/core';
+// import { yellow } from 'chalk';
 
-export function renderActions(actions: Notification[], args: FormatterArgs): void {
-  if (actions && actions.length > 0) {
-    args.writer.categoryHeader('Actions');
-    actions.forEach((action: Notification) => {
-      args.writer.log(`${yellow('■')} ${action.message.text}`);
-    });
-    args.writer.blankLine();
-  }
-}
+// export function renderActions(actions: Notification[], args: FormatterArgs): void {
+//   if (actions && actions.length > 0) {
+//     args.writer.categoryHeader('Actions');
+//     actions.forEach((action: Notification) => {
+//       args.writer.log(`${yellow('■')} ${action.message.text}`);
+//     });
+//     args.writer.blankLine();
+//   }
+// }
 
-export function renderInfo(metadata: CheckupMetadata, args: FormatterArgs) {
-  let { analyzedFilesCount, project } = metadata;
-  let { name, version, repository } = project;
+// export function renderInfo(metadata: CheckupMetadata, args: FormatterArgs) {
+//   let { analyzedFilesCount, project } = metadata;
+//   let { name, version, repository } = project;
 
-  let analyzedFilesMessage =
-    repository.totalFiles !== analyzedFilesCount
-      ? ` (${args.writer.emphasize(`${analyzedFilesCount} files`)} analyzed)`
-      : '';
+//   let analyzedFilesMessage =
+//     repository.totalFiles !== analyzedFilesCount
+//       ? ` (${args.writer.emphasize(`${analyzedFilesCount} files`)} analyzed)`
+//       : '';
 
-  args.writer.blankLine();
-  args.writer.log(
-    `Checkup report generated for ${args.writer.emphasize(
-      `${name} v${version}`
-    )}${analyzedFilesMessage}`
-  );
-  args.writer.blankLine();
-  args.writer.log(
-    `This project is ${args.writer.emphasize(
-      `${repository.age} old`
-    )}, with ${args.writer.emphasize(
-      `${repository.activeDays} active days`
-    )}, ${args.writer.emphasize(`${repository.totalCommits} commits`)} and ${args.writer.emphasize(
-      `${repository.totalFiles} files`
-    )}.`
-  );
-  args.writer.blankLine();
-}
+//   args.writer.blankLine();
+//   args.writer.log(
+//     `Checkup report generated for ${args.writer.emphasize(
+//       `${name} v${version}`
+//     )}${analyzedFilesMessage}`
+//   );
+//   args.writer.blankLine();
+//   args.writer.log(
+//     `This project is ${args.writer.emphasize(
+//       `${repository.age} old`
+//     )}, with ${args.writer.emphasize(
+//       `${repository.activeDays} active days`
+//     )}, ${args.writer.emphasize(`${repository.totalCommits} commits`)} and ${args.writer.emphasize(
+//       `${repository.totalFiles} files`
+//     )}.`
+//   );
+//   args.writer.blankLine();
+// }
 
-export function renderLinesOfCode(info: CheckupMetadata, args: FormatterArgs) {
-  let { repository } = info.project;
+// export function renderLinesOfCode(info: CheckupMetadata, args: FormatterArgs) {
+//   let { repository } = info.project;
 
-  args.writer.sectionedBar(
-    repository.linesOfCode.types.map((type) => {
-      return { title: type.extension, count: type.total };
-    }),
-    repository.linesOfCode.total,
-    'lines of code'
-  );
+//   args.writer.sectionedBar(
+//     repository.linesOfCode.types.map((type) => {
+//       return { title: type.extension, count: type.total };
+//     }),
+//     repository.linesOfCode.total,
+//     'lines of code'
+//   );
 
-  args.writer.blankLine();
-}
+//   args.writer.blankLine();
+// }
 
-export function renderCLIInfo(metadata: CheckupMetadata, args: FormatterArgs) {
-  let { version, configHash } = metadata.cli;
+// export function renderCLIInfo(metadata: CheckupMetadata, args: FormatterArgs) {
+//   let { version, configHash } = metadata.cli;
 
-  args.writer.dimmed(`checkup v${version}`);
-  args.writer.dimmed(`config ${configHash}`);
-  args.writer.blankLine();
-}
+//   args.writer.dimmed(`checkup v${version}`);
+//   args.writer.dimmed(`config ${configHash}`);
+//   args.writer.blankLine();
+// }
