+++ packages/cli/src/formatters/json.ts
@@ -1,11 +1,14 @@
-import { FormatterArgs } from '@checkup/core';
+import { ConsoleWriter, Formatter, FormatterArgs } from '@checkup/core';
 import { writeResultFile } from './file-writer';
 
-export default class JsonFormatter {
+export default class JsonFormatter implements Formatter {
   args: FormatterArgs;
+  writer: ConsoleWriter;
 
   constructor(options: FormatterArgs) {
     this.args = options;
+
+    this.writer = new ConsoleWriter();
   }
 
   format() {
@@ -14,7 +17,7 @@ export default class JsonFormatter {
     if (this.args.outputFile) {
       writeResultFile(log, this.args.cwd, this.args.outputFile);
     } else {
-      this.args.writer.styledJSON(log);
+      this.writer.styledJSON(log);
     }
   }
 }
