+++ packages/core/src/utils/pretty-formatter-utils.ts
@@ -1,6 +1,6 @@
 import { Result } from 'sarif';
+import { BaseOutputWriter } from '..';
 import { groupDataByField } from '../data/formatters';
-import { FormatterArgs } from '../types/cli';
 import { reduceResults, sumOccurrences } from './sarif-utils';
 
 export function renderEmptyResult(taskResult: Result) {
@@ -10,20 +10,18 @@ export function renderEmptyResult(taskResult: Result) {
   };
 }
 
-export function renderLintingSummaryResult(taskResults: Result[], args: FormatterArgs) {
+export function renderLintingSummaryResult(taskResults: Result[], writer: BaseOutputWriter) {
   let groupedTaskResultsByType = groupDataByField(taskResults, 'properties.type');
 
-  args.writer.section(taskResults[0].properties?.taskDisplayName, () => {
+  writer.section(taskResults[0].properties?.taskDisplayName, () => {
     groupedTaskResultsByType.forEach((resultGroup: Result[]) => {
       let groupedTaskResultsByLintRule = reduceResults(
         groupDataByField(resultGroup, 'properties.lintRuleId')
       ).sort((a, b) => (b.occurrenceCount || 0) - (a.occurrenceCount || 0));
       let totalCount = sumOccurrences(groupedTaskResultsByLintRule);
       if (totalCount) {
-        args.writer.subHeader(
-          `${groupedTaskResultsByLintRule[0].properties?.type}s: (${totalCount})`
-        );
-        args.writer.valuesList(
+        writer.subHeader(`${groupedTaskResultsByLintRule[0].properties?.type}s: (${totalCount})`);
+        writer.valuesList(
           groupedTaskResultsByLintRule.map((result) => {
             return result.message.text === 'No results found'
               ? renderEmptyResult(result)
@@ -33,7 +31,7 @@ export function renderLintingSummaryResult(taskResults: Result[], args: Formatte
                 };
           })
         );
-        args.writer.blankLine();
+        writer.blankLine();
       }
     });
   });
